remove old router, app, main and updated root to include links
