--- app/App.tsx
@@ -1,8 +1,0 @@
-import Router from './Router';
-import './styles/globals.css';
-
-const App = () => {
-  return <Router />;
-};
-
-export default App;
