+++ app/root.tsx
@@ -1,7 +1,8 @@
-import { Outlet, Scripts } from '@remix-run/react';
+import { Links, Outlet, Scripts } from '@remix-run/react';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { Toaster } from './components/ui/Toaster';
 import queryClient from './services/queryClient';
+import './styles/globals.css';
 
 const Root = () => {
   return (
@@ -10,6 +11,7 @@ const Root = () => {
         <link rel="icon" type="image/svg+xml" href="/vite.svg" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Vite + React + TS</title>
+        <Links />
       </head>
       <body className="h-screen w-full">
         <div id="root">
